[mod] move group_engines_in_tab to searx.webutils
